requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.3.2
Sphinx==1.3.1